pyproject.toml
[build-system]
requires = ["setuptools>=68", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "memory-system"
version = "0.1.0"
description = "Two-layer AI agent memory system: episodic PostgreSQL log + semantic Obsidian graph"
requires-python = ">=3.11"
dependencies = [
    # Database
    "psycopg[binary]>=3.1",
    "sqlalchemy>=2.0",

    # Web server (ingestion endpoint)
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",

    # HTTP client
    "httpx>=0.27",

    # Config
    "pyyaml>=6.0",
    "python-dotenv>=1.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",

    # LLM clients
    "anthropic>=0.25",
    "openai>=1.23",
    "google-generativeai>=0.5",

    # Telegram
    "python-telegram-bot>=21.0",

    # Utilities
    "rich>=13.7",          # pretty terminal output
    "typer>=0.12",         # CLI
    "croniter>=2.0",       # parse cron expressions for alert rules
    "aiofiles>=23.2",
    "tenacity>=8.2",       # retry logic for API calls
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "ruff>=0.4",
]

[project.scripts]
memory = "ingestion.server:cli_app"

[tool.setuptools.packages.find]
where = ["."]
include = ["ingestion*", "analysis*", "obsidian*", "notifications*", "sync*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
